% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simudata_linear_exp}
\alias{simudata_linear_exp}
\title{Simulated longitudinal dataset: linear vs exponential growth}
\format{
A data frame with samples in rows and time points in columns.
}
\usage{
data(simudata_linear_exp)
}
\description{
A wide data frame where rows are samples and columns are time points (t0, t1, ...),
plus a \code{Group} column indicating the dynamic model ("Linear" or "Exponential").
}
\examples{
data(simudata_linear_exp)
head(simudata_linear_exp)
}
\keyword{datasets}
