% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simudata_periodic_random}
\alias{simudata_periodic_random}
\title{Simulated longitudinal dataset: periodic vs random fluctuations}
\format{
Same structure as \code{simudata_linear_exp}, but groups are
"Periodic" and "Random".
}
\usage{
data(simudata_periodic_random)
}
\description{
Simulated longitudinal dataset: periodic vs random fluctuations
}
\examples{
data(simudata_periodic_random)
head(simudata_periodic_random)
}
\keyword{datasets}
