% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simudata_predator_prey}
\alias{simudata_predator_prey}
\title{Simulated longitudinal dataset: predator–prey dynamics}
\format{
Same structure as \code{simudata_linear_exp}, but groups are
"Predator" and "Prey".
}
\usage{
data(simudata_predator_prey)
}
\description{
Simulated longitudinal dataset: predator–prey dynamics
}
\examples{
data(simudata_predator_prey)
head(simudata_predator_prey)
}
\keyword{datasets}
